+++ x/authz/keeper/keeper_test.go
@@ -4,7 +4,6 @@ import (
 	"testing"
 	"time"
 
-	"github.com/stretchr/testify/require"
 	"github.com/stretchr/testify/suite"
 	tmtime "github.com/tendermint/tendermint/libs/time"
 	tmproto "github.com/tendermint/tendermint/proto/tendermint/types"
@@ -126,8 +125,7 @@ func (s *TestSuite) TestKeeperIter() {
 	granteeAddr := addrs[1]
 	granter2Addr := addrs[2]
 	e := ctx.BlockTime().AddDate(1, 0, 0)
-	sendAuthz, err := banktypes.NewSendAuthorization([]sdk.AccAddress{}, coins100)
-	require.NoError(s.T(), err)
+	sendAuthz := banktypes.NewSendAuthorization(coins100, nil)
 
 	s.authzKeeper.SaveGrant(ctx, granteeAddr, granterAddr, sendAuthz, &e)
 	s.authzKeeper.SaveGrant(ctx, granteeAddr, granter2Addr, sendAuthz, &e)
@@ -147,8 +145,7 @@ func (s *TestSuite) TestDispatchAction() {
 	granterAddr := addrs[0]
 	granteeAddr := addrs[1]
 	recipientAddr := addrs[2]
-	a, err := banktypes.NewSendAuthorization([]sdk.AccAddress{}, coins100)
-	require.NoError(err)
+	a := banktypes.NewSendAuthorization(coins100, nil)
 
 	require.NoError(banktestutil.FundAccount(s.bankKeeper, s.ctx, granterAddr, coins1000))